+++ client/app/_components/Page.jsx
@@ -2,6 +2,7 @@
 
 import { useEffect, useReducer } from 'react'
 import { useRouter } from 'next/navigation'
+import { reducer, initialState } from '@app/state/reducer'
 import { deleteTemplate } from '@/api/'
 import TemplateForm from './forms/TemplateForm'
 import DeleteModal from './modals/DeleteModal'
@@ -9,49 +10,12 @@ import EditModal from './modals/EditModal'
 import PreviewModal from './modals/PreviewModal'
 import Header from './Header'
 import SideBar from './SideBar'
-import Templates from './Templates'
+import Templates from './Templates/Templates'
 import '@szhsin/react-menu/dist/index.css'
 import '@szhsin/react-menu/dist/transitions/slide.css'
 import { signOut, useSession } from '@app/_components/ExternalComponents'
 
 export default function Page({ templates }) {
-  const initialState = {
-    searchQuery: '',
-    filteredTemplates: templates,
-    currentSortType: null,
-    selectedTemplate: null,
-    userFirstLetter: '',
-    createModalVisible: false,
-    editModalVisible: false,
-    deleteModalVisible: false,
-    previewModalVisible: false,
-  }
-
-  const reducer = (state, action) => {
-    switch (action.type) {
-      case 'SET_SEARCH_QUERY':
-        return { ...state, searchQuery: action.payload }
-      case 'SET_FILTERED_TEMPLATES':
-        return { ...state, filteredTemplates: action.payload }
-      case 'SET_CURRENT_SORT_TYPE':
-        return { ...state, currentSortType: action.payload }
-      case 'SET_SELECTED_TEMPLATE':
-        return { ...state, selectedTemplate: action.payload }
-      case 'SET_USER_FIRST_LETTER':
-        return { ...state, userFirstLetter: action.payload }
-      case 'SET_CREATE_MODAL_VISIBLE':
-        return { ...state, createModalVisible: action.payload }
-      case 'SET_EDIT_MODAL_VISIBLE':
-        return { ...state, editModalVisible: action.payload }
-      case 'SET_DELETE_MODAL_VISIBLE':
-        return { ...state, deleteModalVisible: action.payload }
-      case 'SET_PREVIEW_MODAL_VISIBLE':
-        return { ...state, previewModalVisible: action.payload }
-      default:
-        return state
-    }
-  }
-
   const [state, dispatch] = useReducer(reducer, initialState)
   const {
     searchQuery,
